backend/requirements.txt
# ==============================================================
# requirements.txt  —  Backend Dependencies
# ==============================================================

# ── Core Framework ─────────────────────────────────────────────
fastapi==0.135.2
uvicorn==0.42.0
gunicorn==21.2.0
python-multipart==0.0.9

# ── LangChain Stack ────────────────────────────────────────────
langchain==0.3.25
langchain-core==0.3.59
langchain-community==0.3.23
langchain-google-genai==2.1.4
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langchain-groq==0.3.2

# ── LangSmith / LangGraph ──────────────────────────────────────
langsmith==0.3.45
langgraph==0.3.9
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.70

# ── Google AI / GenAI ──────────────────────────────────────────
google-genai==1.68.0
google-auth==2.49.1



# ── OpenAI ─────────────────────────────────────────────────────
openai==1.82.0

# ── Embeddings & Vector Operations ─────────────────────────────
pinecone==7.3.0
langchain-pinecone==0.2.9
sentence-transformers==5.3.0

# ── File Parsing (Resume PDFs / DOCX) ──────────────────────────
PyMuPDF==1.24.10
python-docx==1.1.2
pdfplumber==0.11.4

# ── Utilities ──────────────────────────────────────────────────
pydantic==2.12.5
python-dotenv==1.0.1
tqdm==4.66.5
numpy==2.4.3
pandas==3.0.1
orjson==3.11.7
requests==2.32.5
httpx==0.28.1
httpx-sse==0.4.3
websockets==15.0.1

# ── Text Processing ────────────────────────────────────────────
nltk==3.9.1
spacy==3.8.11

# ── Web Scraping ───────────────────────────────────────────────
beautifulsoup4==4.12.3
lxml==5.1.0
aiohttp==3.10.5

# ── JWT Authentication ─────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
passlib[bcrypt]==1.7.4

supabase==2.28.3